separate components into single files
